refactor: cleanup & move relic sim out of test file
