+++ src/lib/characterPreview/summary/StatScoringSummary.tsx
@@ -4,7 +4,6 @@ import { EnrichedRelics, enrichRelicAnalysis, flatReduction, hashEstTbpRun, Reli
 import { CHARACTER_SCORE } from 'lib/constants/constants'
 import { iconSize } from 'lib/constants/constantsUi'
 import { SingleRelicByPart } from 'lib/gpu/webgpuTypes'
-import { debugEstTbp } from 'lib/relics/estTbp/estTbp'
 import { Assets } from 'lib/rendering/assets'
 import { SimulationScore } from 'lib/scoring/simScoringUtils'
 import DB from 'lib/state/db'
@@ -350,5 +349,3 @@ function RollLine(props: { index: number; relicAnalysis: RelicAnalysis }) {
     </Flex>
   )
 }
-
-debugEstTbp()
